be/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
openpyxl==3.1.2
pillow==12.1.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
requests==2.32.3
